refactor EventCreateRequestDto.java

+++ src/main/java/org/das/event_manager/dto/EventCreateRequestDto.java
@@ -12,6 +12,6 @@ public record EventCreateRequestDto(
         LocalDateTime date,
         BigDecimal cost,
         Integer duration,
-        Integer locationId
+        Long locationId
 ) {
 }
